+++ src/test/java/wooteco/subway/dao/StationDaoTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
+import wooteco.subway.dao.jdbc.JdbcStationDao;
 import wooteco.subway.domain.Station;
 
 import java.util.List;
