+++ src/test/java/wooteco/subway/service/LineServiceTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import wooteco.subway.dao.JdbcLineDao;
+import wooteco.subway.dao.jdbc.JdbcLineDao;
 import wooteco.subway.service.dto.line.LineRequestDTO;
 import wooteco.subway.service.dto.line.LineResponseDTO;
